finished adding product to db minor fixes
