+++ src/main/java/ecommerce/controllers/GetAllProductsController.java
@@ -1,8 +1,8 @@
 package ecommerce.controllers;
 
 import ecommerce.lib.Injector;
-import ecommerce.model.User;
-import ecommerce.service.interfaces.UserService;
+import ecommerce.model.Product;
+import ecommerce.service.interfaces.ProductService;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.ServletException;
@@ -11,17 +11,18 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/user/all")
-public class UsersController extends HttpServlet {
+@WebServlet("/product/all")
+public class GetAllProductsController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("ecommerce");
-    private final UserService userService = (UserService) injector.getInstance(UserService.class);
+    private final ProductService productService
+            = (ProductService) injector.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<User> allUsers = userService.getAll();
-
-        req.setAttribute("users", allUsers);
-        req.getRequestDispatcher("/WEB-INF/views/user/all.jsp").forward(req, resp);
+        List<Product> allProducts = productService.getAll();
+        req.setAttribute("products", allProducts);
+        req.getRequestDispatcher("/WEB-INF/views/product/all.jsp").forward(req, resp);
     }
+
 }
