+++ src/main/java/ecommerce/controllers/InjectDataController.java
@@ -4,15 +4,14 @@ import ecommerce.lib.Injector;
 import ecommerce.model.Product;
 import ecommerce.service.interfaces.ProductService;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/product/all")
-public class ProductsController extends HttpServlet {
+@WebServlet("/inject")
+public class InjectDataController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("ecommerce");
     private final ProductService productService
             = (ProductService) injector.getInstance(ProductService.class);
@@ -20,9 +19,13 @@ public class ProductsController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Product> allProducts = productService.getAll();
-        req.setAttribute("products", allProducts);
-        req.getRequestDispatcher("/WEB-INF/views/product/all.jsp").forward(req, resp);
-    }
+        productService.create(
+                new Product("iPhone1", 999.99));
+        productService.create(
+                new Product("iPhone2", 1999.99));
+        productService.create(
+                new Product("iPhone3", 2999.99));
 
+        req.getRequestDispatcher("/WEB-INF/views/filler.jsp").forward(req, resp);
+    }
 }
