+++ src/main/webapp/WEB-INF/views/cart/shoppingCart.jsp
@@ -26,7 +26,7 @@
     </c:forEach>
 </table>
 <a href="${pageContext.request.contextPath}/registration">Sign up</a><br>
-<a href="${pageContext.request.contextPath}/user/all">All user</a><br>
+<a href="${pageContext.request.contextPath}/user/all">All users</a><br>
 <a href="${pageContext.request.contextPath}/product/all">All products</a><br>
 <a href="${pageContext.request.contextPath}/">Main page</a><br>
 </body>
