+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -5,8 +5,8 @@
 </head>
 <body>
 <a href="${pageContext.request.contextPath}/registration">Sign up</a><br>
-<a href="${pageContext.request.contextPath}/user/all">All user</a><br>
+<a href="${pageContext.request.contextPath}/user/all">All users</a><br>
 <a href="${pageContext.request.contextPath}/product/all">All products</a><br>
-<a href="${pageContext.request.contextPath}/filler">Put some products to DB</a><br>
+<a href="${pageContext.request.contextPath}/inject">Put some products to DB</a><br>
 </body>
 </html>
