+++ src/main/webapp/WEB-INF/views/user/all.jsp
@@ -31,8 +31,6 @@
         </tr>
     </c:forEach>
 </table>
-<a href="${pageContext.request.contextPath}/registration">Sign up</a><br>
-<a href="${pageContext.request.contextPath}>Main page</a><br>
-<a href="${pageContext.request.contextPath}/product/all">All products</a><br>
+<a href="${pageContext.request.contextPath}/">Main page</a><br>
 </body>
 </html>
